+++ server/src/main/java/com/ct/hetingzzz/domain/LoginUser.java
@@ -5,15 +5,15 @@ import java.io.Serializable;
 public class LoginUser implements Serializable {
 
     private static final long serialVersionUID = 288232946557523962L;
-    private Long userId;
+    private Integer userId;
     private String name;
-    private Long roleId;
+    private Integer roleId;
 
-    public Long getUserId() {
+    public Integer getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(Integer userId) {
         this.userId = userId;
     }
 
@@ -25,11 +25,11 @@ public class LoginUser implements Serializable {
         this.name = name;
     }
 
-    public Long getRoleId() {
+    public Integer getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Long roleId) {
+    public void setRoleId(Integer roleId) {
         this.roleId = roleId;
     }
 }
